This was my fix for the problems I had
I created 3 new models. These were ContactClientForm, the SignUp form for employee accounts, and the Testimony submit form. I then created a database that the entered information would be stored in. I allowed only the Employees with their own accounts to be able to access the database information and incorporated editing, working on publishing, as well as login authentication. I created views for the database, views for the forms, and controllers for each database table that was directly correlated with the model.